Added a chrome.storage.onChanged listener that refreshes any StorageItem that was updated.
This is done by adding a refreshers field to StorageImpl,
which is called with the StorageChanges when there is a change.

Each StorageItem adds a refresher to its StorageImpl.refreshers,
which checks if its StorageKey is in the StorageChanges,
and if it is, calls its own refreshers.

Modified Callables and Callable to be
Callables<T = undefined> and Callable<T = undefined>.
This allows an argument to be passed to each Callable and Callables.call(),
or none if T = undefined.
This does require some trickery with the type system to make this work, though,
but the implementation is sound.