Setup Webpack to compile sandbox.html and PolybiusSandbox.js, which will run all sandboxed code.
Since eval doesn't work in the background script
due to the extension's content security policy (CSP),
I need to run eval in a sandboxed page
and then communicate with the background page using window.postMessage().

I didn't set up the JS/TS code for this yet,
but I set up the Webpack compilation for it.

Also, WebAssembly.compile(), etc. violate the CSP, too, so that's not an option either.