Wrote Sandbox.ts (the sandbox) and SandboxMessenger.ts (the client communicating with the sandbox). They communicate using Window.postMessage() and supposedly allow you to use eval, which is unsafe, b/c it's in a sandbox, which has no access to Chrome APIs.
SandboxMessenger exposes to methods: evaluate() and compile().

evaluate(string) evaluates a string directly as JS
and async returns its value.

compile(string) evaluates a string that should be a function
and async returns a function proxy to it.
The actual function cannot be serialized over postMessage(),
but a reference via a unique ID is maintained,
so that when the function proxy is called with arguments (or bind() is called),
it sends the arguments to the sandbox via postMessage() where the function is called,
and then the result is return async to the client.

It all works when I run it not as a Chrome extension,
but even though eval() (really new Function()) is being called in the safe sandbox,
it's still throwing an error that I'm violating the CSP,
even though I specified the sandbox's CSP should allow unsafe eval.

I'm not sure how to fix this.