Wrote a new Collections API based on Java's and with map-like functions like in JS.
It uses inheritance for its interfaces, but its implementations are composition-based.
All methods are bound.

All Collections optionally require a HashEquals<T, H = number | string>,
which hashes an element of type T to type H, which defaults to number or string.
It defaults to HashEquals.default(), which uses referential equality and JSON.stringify to hash.

So far, only ArrayStack, HashMap, and HashSet are implemented.

Collection.basedOn() is used for composition "inheritance".

Ran into some nasty problems with the TypeScript type system:
    No higher kinded types, so a generic type parameter can't be generic itself.
    No contra/covariance, only bivariance in generics.
    keyof doesn't include well known symbols (e.x. Symbol.iterator),
       meaning Difference<T, U> = U \ T (set difference) doesn't fully work.